+++ src/CovidApp.tsx
@@ -59,6 +59,7 @@ import { PrivacyPolicyUSScreen } from '@covid/features/register/us/PrivacyPolicy
 import TermsOfUseUSScreen from '@covid/features/register/us/TermsOfUseUSScreen';
 import i18n from '@covid/locale/i18n';
 import { EditProfileScreen } from '@covid/features/multi-profile/EditProfileScreen';
+import { ArchiveReasonScreen } from '@covid/features/multi-profile/ArchiveReasonScreen';
 
 const Stack = createStackNavigator<ScreenParamList>();
 const Drawer = createDrawerNavigator();
@@ -207,6 +208,7 @@ export default class CovidApp extends Component<object, State> {
         <Stack.Screen name="Login" component={LoginScreen} options={noHeader} />
         <Stack.Screen name="CreateProfile" component={CreateProfileScreen} options={noHeader} />
         <Stack.Screen name="EditProfile" component={EditProfileScreen} options={noHeader} />
+        <Stack.Screen name="ArchiveReason" component={ArchiveReasonScreen} options={noHeader} />
         <Stack.Screen name="ConsentForOther" component={ConsentForOther} options={noHeader} />
         <Stack.Screen name="ReportForOther" component={ReportForOtherScreen} options={noHeader} />
         <Stack.Screen name="SelectProfile" component={SelectProfileScreen} options={noHeader} />
